+++ android/src/main/java/dev/barooni/capacitor/calendar/CapacitorCalendarPlugin.kt
@@ -194,8 +194,20 @@ class CapacitorCalendarPlugin : Plugin() {
             val startDate = call.getLong("startDate")
             val endDate = call.getLong("endDate")
             val isAllDay = call.getBoolean("isAllDay", false)
+            val alertOffsetInMinutes = call.getFloat("alertOffsetInMinutes")
 
-            val eventUri = implementation.createEvent(context, title, calendarId, location, startDate, endDate, isAllDay)
+            val eventUri =
+                implementation.createEvent(
+                    context,
+                    title,
+                    calendarId,
+                    location,
+                    startDate,
+                    endDate,
+                    isAllDay,
+                    alertOffsetInMinutes,
+                )
+
             val id = eventUri?.lastPathSegment ?: throw IllegalArgumentException("Failed to insert event into calendar")
             val ret = JSObject()
             ret.put("result", id)